Fix Windows compile warnings and an error
- Some potentially uninitialized values
- Some unreachable code
- Some signed/unsigned comparisons (object materials: converted the
  material argument passed to material functions from int to uchar)
- One place in mkmaze.c was removing the walls for the room 100% of the
  time instead of 80%. This went back to the original commit whose
  message specified 80%, so I think it is just a debug statement that
  didn't get removed.

+++ include/extern.h
@@ -1402,8 +1402,8 @@ extern struct obj *obj_absorb(struct obj **, struct obj **);
 extern struct obj *obj_meld(struct obj **, struct obj **);
 extern void pudding_merge_message(struct obj *, struct obj *);
 extern void init_obj_material(struct obj *);
-extern boolean valid_obj_material(struct obj *, int);
-extern void set_material(struct obj *, int);
+extern boolean valid_obj_material(struct obj *, uchar);
+extern void set_material(struct obj *, uchar);
 extern struct obj *init_dummyobj(struct obj *, short, long);
 
 /* ### mkroom.c ### */
